[api] refactor: profile unfollow requests are now idempotent as expected (#63)
# Purpose

A request to unfollow another profile should always be idempotent, as
intended for the nature of a HTTP DELETE request method.

# Changelog

## Services
- Updated `ProfileInteractionService` method `unfollow` to simply call
the `FollowRepository` `deleteByFollowerIdAndFollowedId` without
performing any additional checks.

## Repositories
- Removed redundant `FollowRepository` method
`deleteAnyFollowIfExistsBetween`, as this was intended for `Profile`
blocking interactions, which have been removed.
- Updated all repository Jdocs for uniformity.

## Testing
- **Unit testing**:
- Removed redundant `ProfileInteractionService` unit tests based around
the `unfollow` method.
- Removed redundant `ProfileInteractionController` unit tests based
around the `unfollow` method.
- **Integration testing**:
- Added `ProfileInteractionController` integration tests to test the
idempotency of the `unfollowById` method.